Move validator utils to separate file
