+++ test-ts/default-strategy.test.ts
@@ -25,27 +25,29 @@ import { MockVote } from "../typechain-types/MockVote";
 import { SpecificGroupStrategy } from "../typechain-types/SpecificGroupStrategy";
 import {
   ADDRESS_ZERO,
-  deregisterValidatorGroup,
-  electGroup,
-  electMockValidatorGroupsAndUpdate,
   getDefaultGroups,
   getOrderedActiveGroups,
   getUnsortedGroups,
   mineToNextEpoch,
   prepareOverflow,
   randomSigner,
-  registerValidatorAndAddToGroupMembers,
-  registerValidatorAndOnlyAffiliateToGroup,
-  registerValidatorGroup,
   REGISTRY_ADDRESS,
-  removeMembersFromGroup,
   resetNetwork,
   revokeElectionOnMockValidatorGroupsAndUpdate,
   updateGroupCeloBasedOnProtocolStCelo,
-  updateGroupSlashingMultiplier,
   updateMaxNumberOfGroups,
 } from "./utils";
 import { OrderedGroup } from "./utils-interfaces";
+import {
+  deregisterValidatorGroup,
+  electGroup,
+  electMockValidatorGroupsAndUpdate,
+  registerValidatorAndAddToGroupMembers,
+  registerValidatorAndOnlyAffiliateToGroup,
+  registerValidatorGroup,
+  removeMembersFromGroup,
+  updateGroupSlashingMultiplier,
+} from "./utils-validators";
 
 after(() => {
   hre.kit.stop();
