+++ test-ts/end-to-end-specific-group-strategy.test.ts
@@ -15,9 +15,7 @@ import { SpecificGroupStrategy } from "../typechain-types/SpecificGroupStrategy"
 import { StakedCelo } from "../typechain-types/StakedCelo";
 import {
   activateAndVoteTest,
-  activateValidators,
   distributeEpochRewards,
-  electMockValidatorGroupsAndUpdate,
   getGroupsOfAllStrategies,
   getRealVsExpectedCeloForGroups,
   LOCKED_GOLD_UNLOCKING_PERIOD,
@@ -25,13 +23,17 @@ import {
   randomSigner,
   rebalanceDefaultGroups,
   rebalanceGroups,
-  registerValidatorAndAddToGroupMembers,
-  registerValidatorGroup,
   resetNetwork,
   revokeElectionOnMockValidatorGroupsAndUpdate,
   timeTravel,
   upgradeToMockGroupHealthE2E,
 } from "./utils";
+import {
+  activateValidators,
+  electMockValidatorGroupsAndUpdate,
+  registerValidatorAndAddToGroupMembers,
+  registerValidatorGroup,
+} from "./utils-validators";
 
 after(() => {
   hre.kit.stop();
