+++ test-ts/end-to-end.test.ts
@@ -12,20 +12,22 @@ import { SpecificGroupStrategy } from "../typechain-types/SpecificGroupStrategy"
 import { StakedCelo } from "../typechain-types/StakedCelo";
 import {
   activateAndVoteTest,
-  activateValidators,
   distributeEpochRewards,
-  electMockValidatorGroupsAndUpdate,
   LOCKED_GOLD_UNLOCKING_PERIOD,
   mineToNextEpoch,
   randomSigner,
   rebalanceDefaultGroups,
   rebalanceGroups,
-  registerValidatorAndAddToGroupMembers,
-  registerValidatorGroup,
   resetNetwork,
   timeTravel,
   upgradeToMockGroupHealthE2E,
 } from "./utils";
+import {
+  activateValidators,
+  electMockValidatorGroupsAndUpdate,
+  registerValidatorAndAddToGroupMembers,
+  registerValidatorGroup,
+} from "./utils-validators";
 
 after(() => {
   hre.kit.stop();
